moved all job scheduler related code into src.scheduler module (#6)
