+++ src/routes/login.py
@@ -1,5 +1,5 @@
 """
-src/routes/index.py
+src/routes/login.py
 
 Login "/login" route for all HTTP methods.
 """
